+++ src/value.rs
@@ -2,7 +2,7 @@
 use alloc::{string::String, vec::Vec};
 use core::convert::Infallible;
 use core::fmt::{self, Display};
-use core::ops::{Deref, DerefMut};
+use core::ops::{Deref, DerefMut, Index, IndexMut};
 
 use crate::parser::{JsonKind, ParseConfig, ParseDelegate, Parser};
 use crate::{Error, JsonNumber, JsonString};
@@ -67,7 +67,8 @@ impl<'a> Value<'a> {
     /// Returns the [`Object`] inside of this value, if this is a
     /// [`Value::Object`].
     #[must_use]
-    pub fn as_object(&self) -> Option<&Object<'a>> {
+    #[inline]
+    pub const fn as_object(&self) -> Option<&Object<'a>> {
         if let Self::Object(obj) = self {
             Some(obj)
         } else {
@@ -75,10 +76,65 @@ impl<'a> Value<'a> {
         }
     }
 
+    /// Returns a mutable reference to the [`Object`] inside of this value, if
+    /// this is a [`Value::Object`].
+    #[must_use]
+    #[inline]
+    pub fn as_object_mut(&mut self) -> Option<&mut Object<'a>> {
+        if let Self::Object(obj) = self {
+            Some(obj)
+        } else {
+            None
+        }
+    }
+
+    /// Returns the contained value associated with `key`, if this is a
+    /// [`Value::Object`]. Returns `None` if the value is not an object or if
+    /// the key is not found.
+    ///
+    /// # Performance
+    ///
+    /// [`Object`] uses a `Vec` of [`Entry`] types to store its entries. If the
+    /// operation being performed can be done with a single iteration over the
+    /// value's contents instead of multiple random accesses, the iteration
+    /// should be preferred. Additional options to make random access faster in
+    /// environments that can support it [are being considered][issue] for
+    /// future releases.
+    ///
+    /// [issue]: https://github.com/khonsulabs/justjson/issues/7
+    #[must_use]
+    #[inline]
+    pub fn get(&self, key: &str) -> Option<&Value<'a>> {
+        let object = self.as_object()?;
+        object.get(key)
+    }
+
+    /// Returns a mutable reference to the contained value associated with
+    /// `key`, if this is a [`Value::Object`]. Returns `None` if the value is
+    /// not an object or if the key is not found.
+    ///
+    /// # Performance
+    ///
+    /// [`Object`] uses a `Vec` of [`Entry`] types to store its entries. If the
+    /// operation being performed can be done with a single iteration over the
+    /// value's contents instead of multiple random accesses, the iteration
+    /// should be preferred. Additional options to make random access faster in
+    /// environments that can support it [are being considered][issue] for
+    /// future releases.
+    ///
+    /// [issue]: https://github.com/khonsulabs/justjson/issues/7
+    #[must_use]
+    #[inline]
+    pub fn get_mut(&mut self, key: &str) -> Option<&mut Value<'a>> {
+        let object = self.as_object_mut()?;
+        object.get_mut(key)
+    }
+
     /// Returns the [`JsonString`] inside of this value, if this is a
     /// [`Value::String`].
     #[must_use]
-    pub fn as_string(&self) -> Option<&JsonString<'a>> {
+    #[inline]
+    pub const fn as_string(&self) -> Option<&JsonString<'a>> {
         if let Self::String(obj) = self {
             Some(obj)
         } else {
@@ -89,7 +145,8 @@ impl<'a> Value<'a> {
     /// Returns the [`JsonNumber`] inside of this value, if this is a
     /// [`Value::Number`].
     #[must_use]
-    pub fn as_number(&self) -> Option<&JsonNumber<'a>> {
+    #[inline]
+    pub const fn as_number(&self) -> Option<&JsonNumber<'a>> {
         if let Self::Number(obj) = self {
             Some(obj)
         } else {
@@ -100,7 +157,8 @@ impl<'a> Value<'a> {
     /// Returns the `bool` inside of this value, if this is a
     /// [`Value::Boolean`].
     #[must_use]
-    pub fn as_bool(&self) -> Option<bool> {
+    #[inline]
+    pub const fn as_bool(&self) -> Option<bool> {
         if let Self::Boolean(value) = self {
             Some(*value)
         } else {
@@ -111,6 +169,7 @@ impl<'a> Value<'a> {
     /// Returns the slice of values inside of this value, if this is a
     /// [`Value::Array`].
     #[must_use]
+    #[inline]
     pub fn as_array(&self) -> Option<&[Self]> {
         if let Self::Array(value) = self {
             Some(value)
@@ -119,9 +178,42 @@ impl<'a> Value<'a> {
         }
     }
 
+    /// Returns a mutable reference to the Vec of values inside of this value,
+    /// if this is a [`Value::Array`].
+    #[must_use]
+    #[inline]
+    pub fn as_array_mut(&mut self) -> Option<&mut Vec<Self>> {
+        if let Self::Array(value) = self {
+            Some(value)
+        } else {
+            None
+        }
+    }
+
+    /// Returns the contained value at `index`, if this is a [`Value::Array`].
+    /// Returns `None` if the value is not an array or if `index` is beyond the
+    /// bounds of the array.
+    #[must_use]
+    #[inline]
+    pub fn get_index(&self, index: usize) -> Option<&Value<'a>> {
+        let sequence = self.as_array()?;
+        sequence.get(index)
+    }
+
+    /// Returns a mutable reference to the contained value at `index`, if this
+    /// is a [`Value::Array`]. Returns `None` if the value is not an array or if
+    /// `index` is beyond the bounds of the array.
+    #[must_use]
+    #[inline]
+    pub fn get_index_mut(&mut self, index: usize) -> Option<&mut Value<'a>> {
+        let sequence = self.as_array_mut()?;
+        sequence.get_mut(index)
+    }
+
     /// Returns true if this value is `null`/[`Value::Null`].
     #[must_use]
-    pub fn is_null(&self) -> bool {
+    #[inline]
+    pub const fn is_null(&self) -> bool {
         matches!(self, Value::Null)
     }
 
@@ -254,7 +346,7 @@ impl<'a> Value<'a> {
 
 #[test]
 fn value_ases() {
-    assert!(Value::Boolean(true).as_bool().unwrap());
+    assert!(Value::from(true).as_bool().unwrap());
     assert_eq!(
         Value::String(JsonString::from_json("\"\"").unwrap())
             .as_string()
@@ -276,7 +368,7 @@ fn value_ases() {
     assert_eq!(Value::Array(Vec::new()).as_array().unwrap(), &[]);
 
     assert!(Value::Null.is_null());
-    assert!(!Value::Boolean(true).is_null());
+    assert!(!Value::from(true).is_null());
     assert_eq!(Value::Null.as_bool(), None);
     assert_eq!(Value::Null.as_number(), None);
     assert_eq!(Value::Null.as_string(), None);
@@ -485,11 +577,161 @@ where
     }
 }
 
+impl<'a> Index<usize> for Value<'a> {
+    type Output = Value<'a>;
+
+    /// Returns the contained value at `index`, if this is a [`Value::Array`].
+    ///
+    /// # Panics
+    ///
+    /// This function panics if the value is not a [`Value::Array`] or if the
+    /// index is out of bounds of the array.
+    #[inline]
+    fn index(&self, index: usize) -> &Self::Output {
+        self.get_index(index).expect("index not found")
+    }
+}
+
+impl<'a> IndexMut<usize> for Value<'a> {
+    /// Returns a mutable reference to the contained value at `index`, if this
+    /// is a [`Value::Array`].
+    ///
+    /// # Panics
+    ///
+    /// This function panics if the value is not a [`Value::Array`] or if the
+    /// index is out of bounds of the array.
+    #[inline]
+    fn index_mut(&mut self, index: usize) -> &mut Self::Output {
+        self.get_index_mut(index).expect("index not found")
+    }
+}
+
+impl<'b, 'a> Index<&'b str> for Value<'a> {
+    type Output = Value<'a>;
+
+    /// Returns the contained value associated with `key`, if this is a
+    /// [`Value::Object`].
+    ///
+    /// # Panics
+    ///
+    /// This function panics if this value is not a [`Value::Object`] or if the
+    /// `key` is not found.
+    ///
+    /// # Performance
+    ///
+    /// [`Object`] uses a `Vec` of [`Entry`] types to store its entries. If the
+    /// operation being performed can be done with a single iteration over the
+    /// value's contents instead of multiple random accesses, the iteration
+    /// should be preferred. Additional options to make random access faster in
+    /// environments that can support it [are being considered][issue] for
+    /// future releases.
+    ///
+    /// [issue]: https://github.com/khonsulabs/justjson/issues/7
+    #[inline]
+    fn index(&self, index: &'b str) -> &Self::Output {
+        self.get(index).expect("key not found")
+    }
+}
+
+impl<'b, 'a> IndexMut<&'b str> for Value<'a> {
+    /// Returns a mutable reference to the contained value associated with
+    /// `key`, if this is a [`Value::Object`].
+    ///
+    /// # Panics
+    ///
+    /// This function panics if this value is not a [`Value::Object`] or if the
+    /// `key` is not found.
+    ///
+    /// # Performance
+    ///
+    /// [`Object`] uses a `Vec` of [`Entry`] types to store its entries. If the
+    /// operation being performed can be done with a single iteration over the
+    /// value's contents instead of multiple random accesses, the iteration
+    /// should be preferred. Additional options to make random access faster in
+    /// environments that can support it [are being considered][issue] for
+    /// future releases.
+    ///
+    /// [issue]: https://github.com/khonsulabs/justjson/issues/7
+    #[inline]
+    fn index_mut(&mut self, index: &'b str) -> &mut Self::Output {
+        self.get_mut(index).expect("key not found")
+    }
+}
+
+impl<'a> From<bool> for Value<'a> {
+    #[inline]
+    fn from(value: bool) -> Self {
+        Self::Boolean(value)
+    }
+}
+
+impl<'a> From<Object<'a>> for Value<'a> {
+    #[inline]
+    fn from(value: Object<'a>) -> Self {
+        Self::Object(value)
+    }
+}
+
+impl<'a> From<Vec<Value<'a>>> for Value<'a> {
+    #[inline]
+    fn from(value: Vec<Value<'a>>) -> Self {
+        Self::Array(value)
+    }
+}
+
+impl<'a> From<&'a str> for Value<'a> {
+    #[inline]
+    fn from(value: &'a str) -> Self {
+        Self::from(JsonString::from(value))
+    }
+}
+
+impl<'a> From<String> for Value<'a> {
+    #[inline]
+    fn from(value: String) -> Self {
+        Self::from(JsonString::from(value))
+    }
+}
+
+impl<'a> From<JsonString<'a>> for Value<'a> {
+    #[inline]
+    fn from(value: JsonString<'a>) -> Self {
+        Self::String(value)
+    }
+}
+
 /// A JSON Object (list of key-value pairs).
+///
+/// # Performance
+///
+/// [`Object`] uses a `Vec` of [`Entry`] types to store its entries. If the
+/// operation being performed can be done with a single iteration over the
+/// value's contents instead of multiple random accesses, the iteration should
+/// be preferred. Additional options to make random access faster in
+/// environments that can support it [are being considered][issue] for future
+/// releases.
+///
+/// # Modifying an Object
+///
+/// Because [`Object`] internally is just a `Vec<Entry>`, it implements
+/// `Deref`/`DerefMut` to its internal storage. This means that all of `Vec`'s
+/// methods are available to change the contents of an object.
+///
+/// ```rust
+/// use justjson::{Entry, Object, Value};
+///
+/// let mut obj = Object::new();
+/// obj.push(Entry::new("hello", "world"));
+///
+/// assert_eq!(Value::Object(obj).to_json(), r#"{"hello":"world"}"#)
+/// ```
+///
+/// [issue]: https://github.com/khonsulabs/justjson/issues/7
 #[derive(Debug, Eq, PartialEq)]
 pub struct Object<'a>(Vec<Entry<'a>>);
 
 impl<'a> Default for Object<'a> {
+    #[inline]
     fn default() -> Self {
         Self::new()
     }
@@ -498,6 +740,7 @@ impl<'a> Default for Object<'a> {
 impl<'a> Object<'a> {
     /// Returns an empty object.
     #[must_use]
+    #[inline]
     pub const fn new() -> Self {
         Self(Vec::new())
     }
@@ -505,20 +748,47 @@ impl<'a> Object<'a> {
     /// Returns an empty object that can store up to `capacity` elements without
     /// reallocating.
     #[must_use]
+    #[inline]
     pub fn with_capacity(capacity: usize) -> Self {
         Self(Vec::with_capacity(capacity))
+    }
+
+    /// Returns the value associated with `key`, if found.
+    #[must_use]
+    #[inline]
+    pub fn get(&self, key: &str) -> Option<&Value<'a>> {
+        self.iter()
+            .find_map(|entry| (entry.key == key).then_some(&entry.value))
+    }
+
+    /// Returns a mutable reference to the value associated with `key`, if
+    /// found.
+    #[must_use]
+    #[inline]
+    pub fn get_mut(&mut self, key: &str) -> Option<&mut Value<'a>> {
+        self.get_entry_mut(key).map(|entry| &mut entry.value)
+    }
+
+    /// Returns a mutable reference to the entry associated with `key`, if
+    /// found.
+    #[must_use]
+    #[inline]
+    pub fn get_entry_mut(&mut self, key: &str) -> Option<&mut Entry<'a>> {
+        self.iter_mut().find(|entry| entry.key == key)
     }
 }
 
 impl<'a> Deref for Object<'a> {
     type Target = Vec<Entry<'a>>;
 
+    #[inline]
     fn deref(&self) -> &Self::Target {
         &self.0
     }
 }
 
 impl<'a> DerefMut for Object<'a> {
+    #[inline]
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.0
     }
@@ -538,16 +808,30 @@ impl<'a> FromIterator<Entry<'a>> for Object<'a> {
     }
 }
 
+/// An key-value pair in an [`Object`].
 #[derive(Debug, Eq, PartialEq)]
 pub struct Entry<'a> {
+    /// The key of this entry.
     pub key: JsonString<'a>,
+    /// The value associated with the key.
     pub value: Value<'a>,
+}
+
+impl<'a> Entry<'a> {
+    /// Returns a new entry for the given key and value.
+    #[inline]
+    pub fn new(key: impl Into<JsonString<'a>>, value: impl Into<Value<'a>>) -> Self {
+        Self {
+            key: key.into(),
+            value: value.into(),
+        }
+    }
 }
 
 #[test]
 fn primitive_values() {
-    assert_eq!(Value::from_json("true").unwrap(), Value::Boolean(true));
-    assert_eq!(Value::from_json("false").unwrap(), Value::Boolean(false));
+    assert_eq!(Value::from_json("true").unwrap(), Value::from(true));
+    assert_eq!(Value::from_json("false").unwrap(), Value::from(false));
     assert_eq!(Value::from_json("null").unwrap(), Value::Null);
 }
 
@@ -567,11 +851,11 @@ fn objects() {
     assert_eq!(
         Value::from_json(r#" { "hello" : "world" , "another" : "value" }"#).unwrap(),
         Value::Object(Object::from_iter([
-            (
+            Entry::new(
                 JsonString::from_json(r#""hello""#).unwrap(),
                 Value::String(JsonString::from_json(r#""world""#).unwrap())
             ),
-            (
+            Entry::new(
                 JsonString::from_json(r#""another""#).unwrap(),
                 Value::String(JsonString::from_json(r#""value""#).unwrap())
             )
@@ -583,13 +867,44 @@ fn objects() {
 fn cow() {
     let mut value =
         Value::from_json_bytes(br#"{"a":1,"b":true,"c":"hello","d":[],"e":{}}"#).unwrap();
-    let Value::Object(root) = &mut value else { unreachable!() };
+    value["b"] = Value::from(false);
+    let root = value.as_object_mut().unwrap();
     root[0].key = JsonString::from("newa");
+    root[0].value = JsonString::from("a").into();
     let Value::Array(d_array) = &mut root[3].value else { unreachable!() };
     d_array.push(Value::Null);
+
+    // Replace the newly inserted null (uses IndexMut on the array).
+    value["d"][0] = Value::from(false);
+
     let generated = value.to_json();
     assert_eq!(
         generated,
-        r#"{"newa":1,"b":true,"c":"hello","d":[null],"e":{}}"#
+        r#"{"newa":"a","b":false,"c":"hello","d":[false],"e":{}}"#
+    );
+}
+
+#[test]
+fn index() {
+    let mut value = Value::from_json_bytes(br#"{"b":true,"a":[false]}"#).unwrap();
+    assert_eq!(value["b"], Value::from(true));
+    assert_eq!(value.get_index_mut(0), None);
+    assert_eq!(value["a"][0], Value::from(false));
+    assert_eq!(value["a"].get_mut("a"), None);
+}
+
+#[test]
+fn froms() {
+    assert_eq!(Value::from(true), Value::Boolean(true));
+    assert_eq!(Value::from(Object::new()), Value::Object(Object::new()));
+    assert_eq!(Value::from(Vec::new()), Value::Array(Vec::new()));
+    assert_eq!(
+        Value::from(String::from("a")),
+        Value::String(JsonString::from("a"))
+    );
+    assert_eq!(Value::from("a"), Value::String(JsonString::from("a")));
+    assert_eq!(
+        Value::from(JsonString::from("a")),
+        Value::String(JsonString::from("a"))
     );
 }